Update Linkedinbot.py
This file after running the commands go() and store () will create a sql file that has all the job details. It has scraped 350 easy apply  jobs for the job title python. The operation took about 10-15 min with 5 min wait time between every page on linkedin in a total of 40 pages.