Add files via upload
This file , after the user activate the openbrowser() and then he navigates manually to singin to his linkedin account. It reads from a database linkedin.sqlite and it applies automatically for each job in it leaving a whole of 1 min time frame between each application. Also the user must put his resume in the same directory and call it "Technical Resume.docx".